Separated out GraphBuilderFromVector and SimpleGraphBuilder traits
