+++ src/dachshund/simple_transformer.rs
@@ -14,7 +14,7 @@ use crate::dachshund::coreness::Coreness;
 use crate::dachshund::eigenvector_centrality::EigenvectorCentrality;
 use crate::dachshund::error::CLQResult;
 use crate::dachshund::graph_base::GraphBase;
-use crate::dachshund::graph_builder_base::GraphBuilderBase;
+use crate::dachshund::graph_builder_base::GraphBuilderFromVector;
 use crate::dachshund::id_types::{GraphId, NodeId};
 use crate::dachshund::line_processor::{LineProcessor, LineProcessorBase};
 use crate::dachshund::row::{Row, SimpleEdgeRow};
