+++ src/lib.rs
@@ -43,6 +43,7 @@ pub use dachshund::shortest_paths::ShortestPaths;
 pub use dachshund::simple_directed_graph::SimpleDirectedGraph;
 pub use dachshund::simple_directed_graph_builder::SimpleDirectedGraphBuilder;
 pub use dachshund::simple_graph::SimpleGraph;
+pub use dachshund::simple_graph_builder::SimpleGraphBuilder;
 pub use dachshund::simple_transformer::SimpleTransformer;
 pub use dachshund::simple_undirected_graph::SimpleUndirectedGraph;
 pub use dachshund::simple_undirected_graph_builder::SimpleUndirectedGraphBuilder;
