+++ tests/simple_directed_graph.rs
@@ -7,7 +7,7 @@
 extern crate lib_dachshund;
 use lib_dachshund::dachshund::error::{CLQError, CLQResult};
 use lib_dachshund::dachshund::graph_base::GraphBase;
-use lib_dachshund::dachshund::graph_builder_base::GraphBuilderBase;
+use lib_dachshund::dachshund::graph_builder_base::GraphBuilderFromVector;
 use lib_dachshund::dachshund::simple_directed_graph::SimpleDirectedGraph;
 use lib_dachshund::dachshund::simple_directed_graph_builder::SimpleDirectedGraphBuilder;
 use std::collections::HashSet;
