+++ tests/simple_graph.rs
@@ -12,7 +12,7 @@ use lib_dachshund::dachshund::connected_components::{
 };
 use lib_dachshund::dachshund::coreness::Coreness;
 use lib_dachshund::dachshund::error::{CLQError, CLQResult};
-use lib_dachshund::dachshund::graph_builder_base::GraphBuilderBase;
+use lib_dachshund::dachshund::graph_builder_base::GraphBuilderFromVector;
 use lib_dachshund::dachshund::id_types::NodeId;
 use lib_dachshund::dachshund::input::Input;
 use lib_dachshund::dachshund::output::Output;
